+++ src/renderer/components/group-member-item.vue
@@ -29,6 +29,7 @@ export default {
 
 <style lang="less">
 .group-member-item {
+  border: 1px solid red;
   position: relative;
   display: inline-block;
   width: 230px;
