+++ src/renderer/views/chat/message-boxes/left-message-box.vue
