+++ src/renderer/views/chat/message-boxes/right-message-box.vue
