+++ src/renderer/views/chat/message-types/file-message.vue
