+++ src/renderer/views/chat/message-types/image-message.vue
