+++ src/renderer/views/chat/message-types/system-message.vue
