+++ src/renderer/views/chat/message-types/text-message.vue
