+++ src/renderer/views/grouping/category-item.vue
