+++ src/renderer/views/grouping/grouping-item.vue
