+++ src/renderer/views/grouping/index.vue
@@ -1,0 +1,44 @@
+<template>
+  <keep-alive>
+    <component :is="currentComponent" :id="currentContact.id"></component>
+  </keep-alive>
+</template>
+
+<script>
+import DefaultPage from '@/components/default-page'
+import FriendInfo from './friend-info'
+import GroupInfo from './group-info'
+
+export default {
+  name: 'contact-info',
+  components: {
+    DefaultPage,
+    FriendInfo,
+    GroupInfo
+  },
+  data() {
+    return {
+      currentComponent: 'DefaultPage'
+    }
+  },
+  computed: {
+    currentContact() {
+      return this.$store.state.Contact.currentContact
+    }
+  },
+  watch: {
+    currentContact() {
+      if (this.currentContact.isDefaultPage) {
+        this.currentComponent = 'DefaultPage'
+        return
+      }
+
+      this.currentComponent = this.currentContact.isGroup
+        ? 'GroupInfo'
+        : 'FriendInfo'
+    }
+  }
+}
+</script>
+
+<style lang="less"></style>
