+++ src/renderer/views/menu-bar.vue
