+++ src/renderer/views/search-bar.vue
